+++ extras/samples/src_change.py
@@ -1,76 +1,56 @@
 #!/usr/bin/env python
 
+### Example purdy library code
+#
+# Demonstrates the use of methods on Code to modify source before display
+
+from purdy.actions import Append, Transition, Wait
 from purdy.content import Code
+from purdy.ui import SimpleScreen
 
-print('****** Original')
-code = Code('../display_code/doubles.py')
-print(code.source)
+screen = SimpleScreen(starting_line_number=1)
+box = screen.code_box
 
-print('\n****** Remove double with trim')
-code = Code('../display_code/doubles.py').remove_double_blanks()
-print(code.source)
+SOURCE = """\
+    # Before
+    def show_only_this_function():
+        # This comment is deleted
+        # As is this one
+        
 
+        # Above should be a single blank
+        print()
 
-print('\n****** Remove double with trim, respect whitespace')
-code = Code('../display_code/doubles.py').remove_double_blanks(False)
-print(code.source)
+        # This comment should be replaced
+        # Replace between the dollar signs $replace me$
 
-print('\n****** Remove line 5')
-code = Code('../display_code/doubles.py').remove_lines(5)
-print(code.source)
-
-
-print('\n****** Remove line 7-9')
-code = Code('../display_code/doubles.py').remove_lines(7, 2)
-print(code.source)
-
-
-print('\n****** Replace line 5')
-code = Code('../display_code/doubles.py').replace_line(5, 'def new_content():')
-print(code.source)
+        # Fold the numbers in the following
+        # a
+        # 1
+        # 2
+        # b
+        return
 
 
-print('\n****** Inline replace line 5')
-code = Code('../display_code/doubles.py').inline_replace(5, 5, 'inline_replaced():')
-print(code.source)
-
-
-print('\n****** Insert')
-code = Code('../display_code/doubles.py').insert_line(1, '# this is a new comment')
-print(code.source)
-
-
-print('\n****** Python partial')
-code = Code('../display_code/doubles.py').python_portion('foo')
-print(code.source)
-
-
-print('\n****** Left justify NOP')
-code = Code(text="""\
-def foo():
-    pass
-""")
-code.left_justify()
-print(code.source)
-
-
-print('\n****** Left justify')
-code = Code(text="""\
-    def foo():
+    def should_be_culled():
         pass
-""")
-code.left_justify()
-print(code.source)
+"""
 
+before = Code(text=SOURCE)
+after = Code(text=SOURCE).left_justify()
+after.python_portion('show_only_this_function')
+after.insert_line(1, "# After")
+after.remove_lines(3, 2)
+after.replace_line(8, '    # Was replaced')
+after.inline_replace(9, 41, 'done$')
+after.fold_lines(13, 14)
+after.remove_double_blanks()
 
-print('\n****** Fold')
-code = Code(text="""\
-    def foo():
-        d = [
-            1,
-            2,
-            3,
-        ]
-""")
-code.fold_lines(3, 5)
-print(code.source)
+actions = [
+    Append(box, before),
+    Wait(),
+    Transition(box, after),
+]
+
+if __name__ == '__main__':
+    screen.run(actions)